Moving everthing to index
