+++ packages/gatsby-cli/src/init-starter.ts
@@ -9,7 +9,7 @@ import isValid from "is-valid-path"
 import sysPath from "path"
 import prompts from "prompts"
 import url from "url"
-import { updateSiteMetadata } from "gatsby-core-utils/node"
+import { updateSiteMetadata } from "gatsby-core-utils"
 import report from "./reporter"
 import { getPackageManager, promptPackageManager } from "./util/package-manager"
 import { isTTY } from "./util/is-tty"
