--- packages/gatsby-core-utils/node.d.ts
@@ -1,1 +1,0 @@
-export * from "./dist/node"
