--- packages/gatsby-core-utils/node.js
@@ -1,1 +1,0 @@
-module.exports = require("./dist/node")
