+++ packages/gatsby-core-utils/package.json
@@ -27,9 +27,7 @@
   },
   "files": [
     "dist/",
-    "dist/index.d.ts",
-    "node.js",
-    "node.d.ts"
+    "dist/index.d.ts"
   ],
   "dependencies": {
     "ci-info": "2.0.0",
@@ -37,6 +35,7 @@
     "fs-extra": "^8.1.0",
     "node-object-hash": "^2.0.0",
     "proper-lockfile": "^4.1.1",
+    "tmp": "^0.2.1",
     "xdg-basedir": "^4.0.0"
   },
   "devDependencies": {
