+++ packages/gatsby-core-utils/src/index.ts
@@ -6,3 +6,5 @@ export { getCIName, isCI } from "./ci"
 export { createRequireFromPath } from "./create-require-from-path"
 export { getConfigStore } from "./get-config-store"
 export { getGatsbyVersion } from "./get-gatsby-version"
+export * from "./service-lock"
+export * from "./site-metadata"
