--- packages/gatsby-core-utils/src/node.ts
@@ -1,2 +1,0 @@
-export * from "./service-lock"
-export * from "./site-metadata"
