+++ packages/gatsby/src/utils/develop-proxy.ts
@@ -3,7 +3,7 @@ import https from "https"
 import httpProxy from "http-proxy"
 import path from "path"
 import fs from "fs-extra"
-import { getServices } from "gatsby-core-utils/node"
+import { getServices } from "gatsby-core-utils"
 import st from "st"
 import restartingScreen from "./restarting-screen"
 import { IProgram } from "../commands/types"
